refactored with spring + write dataset to database

+++ src/main/java/com/braincourt/preprocessing/DataWriter.java
@@ -1,6 +1,10 @@
 package com.braincourt.preprocessing;
 
 import com.braincourt.preprocessing.dataobjects.DataObject;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -8,68 +12,72 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.stream.Stream;
 
+@Component
 public class DataWriter {
 
-    private final String writePath;
+    private static Logger LOG = LoggerFactory.getLogger(DataWriter.class);
 
     private int processedFiles = 0;
 
-    public DataWriter(String writePath) {
-        this.writePath = writePath;
+    private String jsonFile;
+
+    public DataWriter(@Value("${filename.json}") String jsonFileName) {
+        this.jsonFile = jsonFileName;
     }
 
-    public void writeObjects(Stream<DataObject> dataObjects, String relativePath) {
-        String writeFolderPath = System.getenv("THESIS_PROCESSED_DATA_DIR");
-        File outputFolder = new File(writeFolderPath + relativePath);
+    public void writeObjects(Stream<DataObject> dataObjects, String destinationDir) {
+        File outputFolder = new File(destinationDir);
 
         try {
-            if (outputFolder.isDirectory()) {
-                File[] contents = outputFolder.listFiles();
-                for (int i = 0; i < contents.length; i++) {
-                    File file = contents[i];
-                    file.delete();
-
-                }
-            }
+            deleteContents(outputFolder);
 
             if (outputFolder.exists() && !outputFolder.delete()) {
-                System.err.println("Failed to delete old directory. Exiting...");
+                LOG.error("Failed to delete old directory. Exiting...");
                 System.exit(1);
-
             }
 
             if (outputFolder.mkdir()) {
-                String filePath = outputFolder.getAbsolutePath() + "/data.json";
-                FileWriter writer = new FileWriter(filePath);
-                BufferedWriter bufferedWriter = new BufferedWriter(writer);
+                String filePath = destinationDir + jsonFile;
+                BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(filePath));
 
-                System.out.println(String.format("Beginning processing of %s folder.", outputFolder));
+                LOG.info(String.format("Beginning processing of %s folder.", outputFolder));
 
-                dataObjects.forEach(dataObject -> {
-
-                    try {
-
-                        bufferedWriter.write(dataObject.toJsonString());
-                        bufferedWriter.newLine();
-
-                    } catch (IOException e) {
-
-                        e.printStackTrace();
-
-                    }
+                dataObjects.forEach(dataObject -> { // TODO remove limit after debugging
+                    writeTo(bufferedWriter, dataObject);
                     if (++processedFiles % 1000 == 0) {
-                        System.out.println(String.format("wrote %d lines.", processedFiles));
+                        LOG.info(String.format("wrote %d lines.", processedFiles));
                     }
-
                 });
-                System.out.println(String.format("Finished writing to %s. Lines written: %d.", filePath, processedFiles));
+                LOG.info(String.format("Finished writing to %s. Lines written: %d.", filePath, processedFiles));
                 bufferedWriter.close();
             }
 
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
 
+    private void writeTo(BufferedWriter bw, DataObject dataObject) {
+        try {
 
+            bw.write(dataObject.toJsonString());
+            bw.newLine();
+
+        } catch (IOException e) {
+
+            e.printStackTrace();
+
+        }
+    }
+
+    private void deleteContents(File outputFolder) {
+        if (outputFolder.isDirectory()) {
+            File[] contents = outputFolder.listFiles();
+            if (contents != null && contents.length > 0) {
+                for (File file : contents) {
+                    file.delete();
+                }
+            }
+        }
     }
 }
